React Study 24
##Studies Learned##

1. Router-Dom
Purpose: Use of router-dom to utilise the purpose of React which is loading a page without having to request back to the server again. Faster loading of pages with the use of components created.
- With v6, instead of using switch, we wil not be using Routes.
- We can se "exact" to ensure that react is loading a component that is specifically equal to the target path.
- In creating links on the page, use "Link to" isninstead of "a href"

2. Creating Costom Hooks
Purpose: Ro create something that is reuseable. If you think a function will be used multiple times, it is recomended to turn it into a costom hook instead.
- Ensure that the hook name has the word "use" in the beginning.
- we are returning data that is from the useState.
- utilizing states that are already available on react.