React Study - TSX
Learned:

1. Children
How to pass data using ReactNode and children.
Will have to used onClick where the onclick event will be triggered form the child component. once triggered, will trigger the onClick event on the parent component.

2. TSX
TSX is type sensitive, ensures all the data has a specific type.

3. Using tunnels
" | " is used like an "or" or "any". Means any of the provided list of stings is acceptible, will not trigger anything or show any changes if not on the choices.
  color?:    | "primary"    | "secondary"    | "success"    | "danger"    | "warning"    | "info"    | "light"    | "dark";

4. Props in TSX
 Will have to create a interface everytime.

5. Using props.
You can explicitly call the specific props without having to do it like this "props.item". You can do it like at the bottom:
{item}: Props